requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
openai>=1.0.0
pydantic>=2.7.0
PyMySQL>=1.1.1
python-dotenv>=1.0.1
